Improve information in statistics.

+++ tests/benchmark/src/UAutomizerFileReader.cpp
@@ -104,11 +104,11 @@ void UAutomizerFileReader::action() const {
     sprintf(exec_command, 
         "./../../bin/axd_interpolator QF_TO %s %u 1000 %s;", 
         file_for_implementation.c_str(), curr_solver, file_statistics);
-    int ret = system(exec_command);
-    if(ret != 0){
+    int ret = WEXITSTATUS(system(exec_command));
+    if(ret != 0 && ret != 1){
       char log_command[1000];
       sprintf(log_command, 
-          "echo \"%s\" \"%u\" %d >> \"%s\"", 
+          "echo File: \"%s\" Solver Code: \"%u\" Exit Code: %d >> \"%s\"", 
           file_for_implementation.c_str(), curr_solver, ret, file_statistics);
       system(log_command);
       system(("rm -rf " + temp_file).c_str());
